@@ -1,21 +1,29 @@
 name: dart_chromecast
-version: 0.2.4
+version: 0.3.0
 description: Cast videos to your ChromeCast device using Dart
 homepage: https://github.com/terrabythia/dart_chromecast
+publish_to: none
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=1.20.0'
 
 dependencies:
-  protobuf: '>=1.0.1 <2.0.0'
-  protoc_plugin: ^19.0.0
-  http: ^0.12.2
-  cli_util: ^0.2.0
-  args: ^1.6.0
-  xml: ^4.1.1
-  logging: ^0.11.4
-  multicast_dns: ^0.2.2
-  observable:
+  flutter:
+    sdk: flutter
+  http: ^0.13.0
+  multicast_dns:
     git:
-      url: https://github.com/topilski/observable.git
+      url: https://github.com/flutter/packages.git
+      path: packages/multicast_dns
       ref: master
+  protobuf: ^2.0.0
+  protoc_plugin: ^20.0.0
+  rxdart: ^0.26.0
+
+
+dev_dependencies:
+  dart_rules:
+    git:
+      url: https://github.com/fastogt/dart_rules.git
+      ref: main
