@@ -3,8 +3,9 @@ import 'dart:io';
 import 'package:dart_chromecast/casting/cast_channel.dart';
 
 class ConnectionChannel extends CastChannel {
-
-  ConnectionChannel.create(Socket socket, {String sourceId, String destinationId}) :
-        super.CreateWithSocket(socket, sourceId: sourceId ?? 'sender-0', destinationId: destinationId ?? 'receiver-0', namespace: 'urn:x-cast:com.google.cast.tp.connection');
-
+  ConnectionChannel.create(Socket? socket, {String? sourceId, String? destinationId})
+      : super.createWithSocket(socket,
+            sourceId: sourceId ?? 'sender-0',
+            destinationId: destinationId ?? 'receiver-0',
+            namespace: 'urn:x-cast:com.google.cast.tp.connection');
 }
