@@ -3,8 +3,9 @@ import 'dart:io';
 import 'package:dart_chromecast/casting/cast_channel.dart';
 
 class MediaChannel extends CastChannel {
-
-  MediaChannel.Create({Socket socket, String sourceId, String destinationId}) :
-      super.CreateWithSocket(socket, sourceId: sourceId, destinationId: destinationId, namespace: 'urn:x-cast:com.google.cast.media');
-
+  MediaChannel.create({Socket? socket, String? sourceId, String? destinationId})
+      : super.createWithSocket(socket,
+            sourceId: sourceId,
+            destinationId: destinationId,
+            namespace: 'urn:x-cast:com.google.cast.media');
 }
